requirements.txt
aiogram==2.25.2
openai
python-dotenv
Pillow==9.5.0
